Made stuff Prettier again

+++ CacheSimulator/src/main/java/pkg/Driver.java
@@ -21,7 +21,7 @@ public class Driver {
 
 			Simulation sim = new Simulation(KN, (int) Math.pow(2,  K), Simulation.FIFO);
 			sim.run(firstTrace);
-			writer.print(Simulation.FIFO + "\tKN: " + KN + "\tK: " + Math.pow(2,  K) + "\t");
+			writer.print(Simulation.FIFO + "\tKN: " + KN + "\tK: " + (int) Math.pow(2,  K) + "\t");
 			writer.println(sim.getMisses() + "\t" + sim.getAdresses() + "\t" + (100f * sim.getMisses() / sim.getAdresses()) + "\n");
 
 			K++;
@@ -39,7 +39,7 @@ public class Driver {
 
 			Simulation sim = new Simulation(KN, (int) Math.pow(2,  K), Simulation.LRU);
 			sim.run(firstTrace);
-			writer.print(Simulation.LRU + "\tKN: " + KN + "\tK: " + Math.pow(2,  K) + "\t");
+			writer.print(Simulation.LRU + "\tKN: " + KN + "\tK: " + (int) Math.pow(2,  K) + "\t");
 			writer.println(sim.getMisses() + "\t" + sim.getAdresses() + "\t" + (100f * sim.getMisses() / sim.getAdresses()) + "\n");
 
 			K++;
@@ -58,7 +58,7 @@ public class Driver {
 
 			Simulation sim = new Simulation(KN, (int) Math.pow(2,  K), Simulation.FIFO);
 			sim.run(secondTrace);
-			writer.print(Simulation.FIFO + "\tKN: " + KN + "\tK: " + Math.pow(2,  K) + "\t");
+			writer.print(Simulation.FIFO + "\tKN: " + KN + "\tK: " + (int) Math.pow(2,  K) + "\t");
 			writer.println(sim.getMisses() + "\t" + sim.getAdresses() + "\t" + (100f * sim.getMisses() / sim.getAdresses()) + "\n");
 
 			K++;
@@ -76,7 +76,7 @@ public class Driver {
 
 			Simulation sim = new Simulation(KN, (int) Math.pow(2,  K), Simulation.LRU);
 			sim.run(secondTrace);
-			writer.print(Simulation.LRU + "\tKN: " + KN + "\tK: " + Math.pow(2,  K) + "\t");
+			writer.print(Simulation.LRU + "\tKN: " + KN + "\tK: " + (int) Math.pow(2,  K) + "\t");
 			writer.println(sim.getMisses() + "\t" + sim.getAdresses() + "\t" + (100f * sim.getMisses() / sim.getAdresses()) + "\n");
 
 			K++;
